requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.9.0
anthropic>=0.21.3
python-dotenv>=1.0.1